Add: src/apis/accountSetting.jsをTypeScript対応しました
