--- frontend/src/apis/accountSetting.js
@@ -1,34 +1,0 @@
-import axios from 'axios';
-import { accountSettings } from '../urls/index.ts';
-
-// アカウントを更新した情報を取得するAPIコール関数
-// postの第3引数にwithCredentials: trueを指定することで、
-// API(Rails)と通信する際にデータにcookieを含めることができる
-export const patchAccountSetting = async ({
-  user: {
-    id,
-    name,
-    email,
-    open_rank,
-  },
-  image
-}) => {
-  try {
-    const response = await axios.patch(
-      accountSettings(id),
-      {
-        user: {
-          name: name,
-          email: email,
-          open_rank: open_rank,
-        },
-        image: image
-      },
-      { withCredentials: true }
-    );
-    axios.defaults.headers.common['X-CSRF-Token'] = response.headers['x-csrf-token'];
-    return response.data;
-  } catch(e) {
-    throw e;
-  }
-};
